Ajout d'une barre de vie, pour le moment elle est racorder sur les piece car il n'y as pas de vie, le fond est noir sois il faut changer la barre en noir et changer la couleur des ecriture sois mettre dans une autre couleur normalement c'est pas dur faut juste choisir la couleur

+++ src/mariaLost/gamePlay/view/GameView.java
@@ -1,6 +1,7 @@
 package mariaLost.gamePlay.view;
 
 import javafx.animation.AnimationTimer;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
@@ -70,7 +71,7 @@ public class GameView extends BorderPane {
         animationTimer.stop();
     }
 
-    public SimpleIntegerProperty lifeProperty() {
+    public SimpleDoubleProperty lifeProperty() {
         return barController.lifeProperty();
     }
